conditonal render for questionnaire in navbar

+++ src/routes/dashboard.js
@@ -1,10 +1,11 @@
 import authAxios from "../lib/authAxios"
 import apiRoute from "../lib/apiRoute"
 import { useEffect, useState } from "react"
-
+import { Link } from "react-router-dom"
+import UploadWidget from "../components/uploadWidget"
 const Dashboard = () => {
 
-//states
+    //states
     const [answers, setAnswers] = useState([])
     const [questions, setQuestions] = useState([])
     const [currentPage, setCurrentPage] = useState(1)
@@ -12,7 +13,7 @@ const Dashboard = () => {
     const [updateOpen, setUpdateOpen] = useState(false)
     const [updatedAnswer, setUpdatedAnswer] = useState()
 
-//async function that initially fetches the first answer and then subsequent answers afterwards
+    //async function that initially fetches the first answer and then subsequent answers afterwards
     
     async function fetchQuestionsAndAnswers() {
         const questionOffset = (currentPage - 1) * perPage;
@@ -28,31 +29,32 @@ const Dashboard = () => {
         let fetchedAnswers;
         if (!updateOpen) {
             fetchedAnswers = answerData.data.data.sort((a, b) => a.question_id - b.question_id).map(answer => (
-                <li className="m-8 flex" key={answer.question_id}><div className="h-64 overflow-scroll">{answer.answer}</div><div><button id={answer.question_id} onClick={() => { setUpdateOpen(true); setUpdatedAnswer(answer.answer)}} className="button text-black my-2 mx-8">Update</button></div></li>
+                <li className="m-8 flex" key={answer.question_id}><div className="h-64 overflow-scroll scrolling">{answer.answer}</div><div><button id={answer.question_id} onClick={() => { setUpdateOpen(true); setUpdatedAnswer(answer.answer) }} className="button text-black my-2 mx-8">Update</button></div></li>
             
             ))
-        // Option that turns answer text into a textarea to edit the answer 
+            // Option that turns answer text into a textarea to edit the answer 
         } else {
-           fetchedAnswers = answerData.data.data.sort((a, b) => a.question_id - b.question_id).map(answer => (
-               <li className="m-8 flex" key={answer.question_id}><form className="overflow-scroll"><textarea id={answer.question_id} className="w-[800px] h-64 bg-black-transparent2 border border-gold rounded-lg box-shadow focus:outline-none p-4 m-2" value={updatedAnswer} onChange={(e) => setUpdatedAnswer(e.target.value)}/></form><div><button id={answer.question_id} onClick={(e) => { saveUpdate(e)}} className="button text-black my-16 mx-16">Save</button></div></li>
-            )); 
+            fetchedAnswers = answerData.data.data.sort((a, b) => a.question_id - b.question_id).map(answer => (
+                <li className="m-8 flex" key={answer.question_id}><form className="overflow-scroll scroll scrolling"><textarea id={answer.question_id} className="w-[800px] h-64 bg-black-transparent2 border border-gold rounded-lg box-shadow focus:outline-none p-4 m-2" value={updatedAnswer} onChange={(e) => setUpdatedAnswer(e.target.value)} /></form><div><button id={answer.question_id} onClick={(e) => { saveUpdate(e) }} className="button text-black my-16 mx-16">Save</button></div></li>
+            ));
            
         }
         console.log(answerData)
         // map that renders questions from DB
         const fetchedQuestions = questionData.data.data.map(question => (
-            <li className="m-8" key={question.id}><div className="h-32 text-lg overflow-scroll"><span className="forum text-3xl">{question.id}.</span><br />{question.text}</div></li>
+            <li className="m-8" key={question.id}><div className="h-32 text-lg overflow-scroll scrolling"><span className="forum text-3xl">{question.id}.</span><br />{question.text}</div></li>
         ));
       
         setQuestions(fetchedQuestions)
         setAnswers(fetchedAnswers)
     }
     
+    console.log(answers)
 
     /// manage pagination for answers
     const nextPage = () => {
         setCurrentPage(currentPage + 1)
-    }  
+    }
     const prevPage = () => {
         if (currentPage > 1) {
             setCurrentPage(currentPage - 1)
@@ -65,11 +67,11 @@ const Dashboard = () => {
     //saves updated answer using patch method
     const saveUpdate = async (e) => {
         try {
-            const response = await authAxios.patch(`${apiRoute}/api/v1/answers/${e.target.id}`, {question_id: e.target.id, answer: updatedAnswer})
+            const response = await authAxios.patch(`${apiRoute}/api/v1/answers/${e.target.id}`, { question_id: e.target.id, answer: updatedAnswer })
             setUpdateOpen(false)
         } catch (error) {
             console.log(error)
-            }
+        }
     }
 
     /// Delete all answers 
@@ -89,28 +91,43 @@ const Dashboard = () => {
     }, [updateOpen, updatedAnswer]);
 
       
-    return (
-        <div className="flex flex-col justify-center items-center">
-            <div className="bg-black-transparent border border-white p-16 rounded-lg w-11/12 overflow-auto">
-                <h1 className="forum text-5xl">Dashboard</h1>
-                <h2 className="forum text-3xl mt-4">Your Answers</h2>
-                <div className="flex flex-col-2">
-                    <ul className="mont w-96">{questions.slice((currentPage - 1) * perPage, currentPage * perPage)}</ul>
-                    <ul className="mont flex-1">{answers.slice((currentPage - 1) * perPage, currentPage * perPage)}</ul>
-                </div>
+    if (answers.length > 0) {
+        return (
+            <div className="flex flex-col justify-center items-center">
+                <div className="bg-black-transparent border border-white p-16 rounded-lg w-11/12 overflow-auto">
+                    <h1 className="forum text-5xl">Dashboard</h1>
+                    <h2 className="forum text-3xl mt-4">Your Answers</h2>
+                    {/* <UploadWidget /> */}
+                    <div className="flex flex-col-2">
+                        <ul className="mont w-96">{questions.slice((currentPage - 1) * perPage, currentPage * perPage)}</ul>
+                        <ul className="mont flex-1 mt-10">{answers.slice((currentPage - 1) * perPage, currentPage * perPage)}</ul>
+                    </div>
                     <div className="flex justify-around mt-4">
                         <button className="button" onClick={prevPage} disabled={isPrevDisabled}>
                             Previous
                         </button>
                         <button className="button" onClick={nextPage} disabled={isNextDisabled}>
                             Next
-                    </button>
+                        </button>
                     
-                </div>
+                    </div>
                 
+                </div>
+                <button className="button mt-8" onClick={deleteAll}>Delete All Your Answers </button>
             </div>
-            <button className="button mt-8" onClick={deleteAll}>Delete All Your Answers </button>
-        </div>
-    )
+       
+
+        )
+    } else {
+        return (
+            <div className="flex flex-col justify-center items-center">
+                <div className="bg-black-transparent border border-white p-16 rounded-lg w-11/12 overflow-auto">
+                    <h1 className="forum text-5xl">Dashboard</h1>
+                    <p className="forum text-2xl mt-8">Welcome to sorcery portal, please fill out our business exploration questionnaire</p>
+                    <Link to={"/questionnaire"}><button className="button mt-16">Start here</button></Link>
+                </div>
+                </div>
+        )
+    }
 }
 export default Dashboard
